# apps/docs/src/data/openapi/nest-auth.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/signup": {
      "post": {
        "description": "Register a new user. Response format depends on accessTokenType configuration:\n- Header mode (default): Returns tokens in response body\n- Cookie mode: Sets tokens in HTTP-only cookies and returns success message",
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthSignupRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Header mode: Returns message + tokens in body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthWithTokensResponseDto"
                }
              }
            }
          }
        },
        "summary": "Signup",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticate user. Response format depends on accessTokenType configuration:\n- Header mode (default): Returns tokens in response body\n- Cookie mode: Sets tokens in HTTP-only cookies and returns success message",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthLoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Header mode: Returns message + tokens in body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthWithTokensResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NestAuthResetPasswordWithTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NestAuthPasswordResetResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "@ackplus/nest-auth API",
    "description": "OpenAPI specification generated from the Nest Auth module",
    "version": "2.0.0-beta.16",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "NestAuthSignupRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address (required if phone not provided)",
            "example": "user@example.com"
          },
          "phone": {
            "type": "string",
            "description": "User phone number (required if email not provided)",
            "example": "+1234567890"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "SecurePass123!",
            "minLength": 8
          },
          "tenantId": {
            "type": "string",
            "description": "Tenant ID for multi-tenant applications",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "guard": {
            "type": "string",
            "description": "Guard context (e.g. admin, web, vendor) for isolation. Deprecated: use client",
            "example": "admin",
            "deprecated": true
          }
        },
        "required": [
          "password"
        ]
      },
      "AuthWithTokensResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token (short-lived)",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxMjMiLCJpYXQiOjE2OTk5OTk5OTksImV4cCI6MTY5OTk5OTk5OX0.xyz"
          },
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token (long-lived)",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxMjMiLCJ0eXBlIjoicmVmcmVzaCIsImlhdCI6MTY5OTk5OTk5OX0.abc"
          },
          "message": {
            "type": "string",
            "description": "Success message (added by controller based on configuration)",
            "example": "Login successful"
          },
          "isRequiresMfa": {
            "type": "boolean",
            "description": "Whether multi-factor authentication is required",
            "example": false
          },
          "mfaMethods": {
            "type": "array",
            "description": "Available MFA methods when isRequiresMfa is true",
            "example": [
              "email",
              "totp"
            ],
            "items": {
              "type": "string",
              "enum": [
                "email",
                "sms",
                "totp"
              ]
            }
          },
          "defaultMfaMethod": {
            "type": "string",
            "description": "Default/recommended MFA method",
            "example": "email",
            "enum": [
              "email",
              "sms",
              "totp"
            ]
          },
          "trustToken": {
            "type": "string",
            "description": "Trust token for trusted device verification",
            "example": "1234567890"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "isRequiresMfa"
        ]
      },
      "EmailCredentialsDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "SecurePass123!",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "PhoneCredentialsDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "User phone number",
            "example": "+1234567890"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "SecurePass123!",
            "minLength": 8
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "SocialCredentialsDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "OAuth token or ID token from social provider",
            "example": "ya29.a0AfH6SMBx1234567890abcdefghijklmnop"
          }
        },
        "required": [
          "token"
        ]
      },
      "PasswordlessOtpCredentialsDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Email or phone (same value as in send request)",
            "example": "user@example.com"
          },
          "channels": {
            "type": "array",
            "description": "Channel(s) to try in order. Use both when the client is unsure whether `identifier` is email or phone.",
            "example": [
              "email",
              "sms"
            ],
            "items": {
              "type": "string",
              "enum": [
                "email",
                "sms"
              ]
            }
          },
          "code": {
            "type": "string",
            "description": "One-time code from email or SMS",
            "example": "123456"
          }
        },
        "required": [
          "identifier",
          "channels",
          "code"
        ]
      },
      "NestAuthLoginRequestDto": {
        "type": "object",
        "properties": {
          "providerName": {
            "type": "string",
            "description": "Authentication provider name",
            "example": "email",
            "enum": [
              "email",
              "phone",
              "passwordless",
              "google",
              "facebook",
              "apple",
              "github"
            ],
            "default": "email"
          },
          "credentials": {
            "description": "Login credentials - type varies by provider",
            "examples": {
              "emailLogin": {
                "summary": "Email + password",
                "value": {
                  "email": "user@example.com",
                  "password": "SecurePass123!"
                }
              },
              "phoneLogin": {
                "summary": "Phone + password",
                "value": {
                  "phone": "+1234567890",
                  "password": "SecurePass123!"
                }
              },
              "passwordlessOtp": {
                "summary": "Passwordless OTP — set providerName to passwordless (after POST /auth/passwordless/send)",
                "value": {
                  "providerName": "passwordless",
                  "credentials": {
                    "identifier": "user@example.com",
                    "channels": [
                      "email",
                      "sms"
                    ],
                    "code": "123456"
                  }
                }
              },
              "socialLogin": {
                "summary": "Social Login (Google/Facebook/etc)",
                "value": {
                  "token": "ya29.a0AfH6SMBx...",
                  "type": "idToken"
                }
              }
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmailCredentialsDto"
              },
              {
                "$ref": "#/components/schemas/PhoneCredentialsDto"
              },
              {
                "$ref": "#/components/schemas/SocialCredentialsDto"
              },
              {
                "$ref": "#/components/schemas/PasswordlessOtpCredentialsDto"
              }
            ]
          },
          "tenantId": {
            "type": "string",
            "description": "Tenant ID for multi-tenant applications",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "guard": {
            "type": "string",
            "description": "Guard context (e.g. admin, web, vendor) for isolation. Deprecated: use client",
            "example": "admin",
            "deprecated": true
          },
          "createUserIfNotExists": {
            "type": "boolean",
            "description": "Auto-create user if not exists (for social auth)",
            "default": false
          }
        },
        "required": [
          "credentials"
        ]
      },
...
      "NestAuthResetPasswordWithTokenRequestDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Password reset token (JWT) received after OTP verification",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0eXBlIjoicGFzc3dvcmQtcmVzZXQifQ.xyz"
          },
          "newPassword": {
            "type": "string",
            "description": "New password",
            "example": "NewSecurePass123!",
            "minLength": 8
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "NestAuthPasswordResetResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Password has been reset successfully"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}