# packages/nest-auth/src/lib/auth/dto/responses/auth.response.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import {
    ITokensResponse,
    IUserResponse,
    IAuthResponse,
    IVerify2faResponse,
    NestAuthMFAMethodEnum,
    INestAuthTenant,
    INestAuthUserAccess,
} from '@ackplus/nest-auth-contracts';

/**
 * Authentication tokens response
 */
export class AuthTokensResponseDto implements ITokensResponse {
    @ApiProperty({
        description: 'JWT access token (short-lived)',
        example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxMjMiLCJpYXQiOjE2OTk5OTk5OTksImV4cCI6MTY5OTk5OTk5OX0.xyz',
    })
    accessToken: string;

    @ApiProperty({
        description: 'JWT refresh token (long-lived)',
        example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxMjMiLCJ0eXBlIjoicmVmcmVzaCIsImlhdCI6MTY5OTk5OTk5OX0.abc',
    })
    refreshToken: string;
}

/**
 * User information response
 */
export class UserResponseDto implements IUserResponse {
    @ApiProperty({
        description: 'User unique identifier',
        example: '123e4567-e89b-12d3-a456-426614174000',
    })
    id: string;

    @ApiPropertyOptional({
        description: 'User email address',
        example: 'user@example.com',
    })
    email?: string;

    @ApiPropertyOptional({
        description: 'User phone number',
        example: '+1234567890',
    })
    phone?: string;

    @ApiProperty({
        description: 'Email verification status',
        example: true,
    })
    emailVerifiedAt: Date;

    @ApiProperty({
        description: 'Phone verification status',
        example: true,
    })
    phoneVerifiedAt: Date;

    @ApiPropertyOptional({
        description: 'Whether MFA is enabled for this user',
        example: false,
    })
    isMfaEnabled?: boolean;

    @ApiPropertyOptional({
        description: 'User roles (role names)',
        example: ['admin', 'user'],
        type: [String],
    })
    roles?: string[];

    @ApiPropertyOptional({
        description: 'User permissions (flattened from roles)',
        example: ['read:users', 'write:users'],
        type: [String],
    })
    permissions?: string[];

    @ApiPropertyOptional({
        description: 'Additional user metadata',
        example: { firstName: 'John', lastName: 'Doe' },
    })
    metadata?: Record<string, any>;

    @ApiPropertyOptional({
        description: 'Tenant ID for multi-tenant applications',
        example: 'tenant-123',
    })
    tenantId?: string;
}

/**
 * Authentication Response with Tokens DTO
 *
 * Used in header mode (accessTokenType: 'header')
 * Returns tokens in the response body along with message and status.
 */
export class AuthWithTokensResponseDto extends AuthTokensResponseDto implements IAuthResponse {
    @ApiPropertyOptional({
        description: 'Success message (added by controller based on configuration)',
        example: 'Login successful',
    })
    message?: string;

    @ApiProperty({
        description: 'Whether multi-factor authentication is required',
        example: false,
    })
    isRequiresMfa: boolean;

    @ApiPropertyOptional({
        description: 'Available MFA methods when isRequiresMfa is true',
        example: ['email', 'totp'],
        enum: NestAuthMFAMethodEnum,
        isArray: true,
    })
    mfaMethods?: NestAuthMFAMethodEnum[];

    @ApiPropertyOptional({
        description: 'Default/recommended MFA method',
        example: 'email',
        enum: NestAuthMFAMethodEnum,
    })
    defaultMfaMethod?: NestAuthMFAMethodEnum;

    @ApiPropertyOptional({
        description: 'Trust token for trusted device verification',
        example: '1234567890',
    })
    trustToken?: string;
}

// Alias for backward compatibility
export type AuthResponseDto = AuthWithTokensResponseDto;

/**
 * Two-factor Authentication Response with Tokens DTO
 *
 * Used in header mode (accessTokenType: 'header')
 * Returns tokens in the response body after successful 2FA verification.
 */
export class Verify2faWithTokensResponseDto extends AuthTokensResponseDto implements IVerify2faResponse {
    @ApiPropertyOptional({
        description: 'Verification success message (added by controller)',
        example: '2FA verification successful',
    })
    message?: string;

    @ApiPropertyOptional({
        description: 'User information with roles and permissions',
        type: UserResponseDto,
    })
    user?: UserResponseDto;
}

// Alias for backward compatibility
export type Verify2faResponseDto = Verify2faWithTokensResponseDto;